pyproject.toml
[tool.poetry]
name = "option-util"
version = "0.1.1"
description = ""
authors = ["David Mckim <davidmckim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.1"
flake8 = "^7.1.1"
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
